installed Icons and added Links file
