+++ src/Components/Footer.css
@@ -1,0 +1,18 @@
+
+.footerContainer {
+    display: flex;
+    flex-direction: column;
+    box-sizing: border-box;
+    height: 80px;
+    width: 100vw; 
+    display: flex;
+    background-repeat: no-repeat;
+    background-size: cover;
+    justify-content: center;
+    align-items: center;
+    position: bottom;
+    bottom:0; left:0; right:0;
+    margin-top: 60vh;
+    box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.25);
+    border-bottom: 1px solid gray;
+}
