+++ src/Components/Footer.js
@@ -1,7 +1,8 @@
 import React from 'react';
+import '../Components/Footer.css'
+
 
 const Footer = () => {
- 
     return (
         <div className= "footerContainer">
             <div className="footer">
